+++ CHANGES.md
@@ -5,4 +5,5 @@
 - Import a default http client to communicate with the Equinix API (#25, #22, #23, @maiste)
 - Add some user commands (#24, @maiste)
 - Add some organization commands (#33, @maiste)
+- Add some projects commands (#36, @maiste)
 - JSON manager module (#6, @maiste)